+++ src/Application/Observability/ObservabilityException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EventFlow\Application\Observability;
+
+use RuntimeException;
+
+final class ObservabilityException extends RuntimeException
+{
+}